requirements.txt
fastapi==0.111.1
scikit-learn==1.4.2
pandas==2.2.2
uvicorn==0.30.1
notebook==7.2.1
nltk==3.8.1
langchain_community==0.2.7
langchain==0.2.9
langchain_groq==0.1.6
langchain_core==0.2.21
llama-parse==0.4.9
python-dotenv==1.0.1
groq==0.9.0
supabase==2.5.3